backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
django-cors-headers==4.6.0
django-simple-history==3.7.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
reportlab==4.2.5
redis==5.2.1
python-decouple==3.8
Pillow==11.1.0
dj-database-url==2.3.0
